requirements.txt
mcp == 1.11.0
dotenv == 0.9.9
langchain-ollama == 0.3.2
langchain-google-genai == 2.1.6
langchain-openai == 0.3.28
mcp_use == 1.3.3
grpcio-tools == 1.73.1
aiofiles == 24.1.0
fastmcp == 2.10.6
black == 25.1.0